Admin links to working homepage.
This is a bit of a hack. The i18n/l10n works in a surprising way, and
scopes the first section of all non-admin URLs with the language:

  /<language>/home
  /en/home

Which means that once you navigate away from one of these localised
pages (e.g. /validations), you lose the scoped URL and any rails paths
are ruined:

  /en/home -> /validations
  links to home_path become /home rather than /en/home
  meaning, the language is set to home and the url redirects to /home/home

I don't have time to fix/work around this, so we have this commit.